// views/Dashboard/dashboardMilestones/types.ts
export interface MilestoneDataset {
  label: string
  data: number[]
  backgroundColor: string[]
  hoverOffset: number
  borderWidth?: number
}

export interface MilestoneChartData {
  labels: string[]
  datasets: MilestoneDataset[]
  id: number
  progress: number
  isActive: boolean
}

export interface MilestoneChartOptions {
  maintainAspectRatio: boolean
  responsive: boolean
  cutout: string
  plugins: {
    title: {
      display: boolean
      text: string
      fontSize: number
      fontColor: string
    }
  }
}

export interface IMilestoneCard {
  data: MilestoneChartData
  options: MilestoneChartOptions
}

export interface IMilestone {
  data: MilestoneChartData[]
}